+++ satella/coding/structures/sorted_list.py
@@ -4,6 +4,8 @@ import collections
 
 logger = logging.getLogger(__name__)
 
+__all__ = ['SortedList', 'SliceableDeque']
+
 T = tp.TypeVar('T')
 
 
@@ -22,9 +24,9 @@ class SortedList(tp.Generic[T]):
 
     def __init__(self, items: tp.Iterable[T] = (), key: tp.Callable[[T], int] = lambda a: a):
         sort = sorted((key(item), item) for item in items)
-        self.items = collections.deque(a[1] for a in sort)
-        self.keys = collections.deque(a[0] for a in sort)
-        self.key = key
+        self.items = SliceableDeque(a[1] for a in sort)     # type: SliceableDeque[T]
+        self.keys = collections.deque(a[0] for a in sort)      # type: collections.deque[int]
+        self.key = key                                      # type: tp.Callable[[T], int]
 
     def __contains__(self, item: T) -> bool:
         return item in self.items
@@ -44,7 +46,7 @@ class SortedList(tp.Generic[T]):
         for elem in elements:
             self.add(elem)
 
-    def __getitem__(self, item: int) -> tp.Union[T, tp.List[T]]:
+    def __getitem__(self, item: tp.Union[slice, int]) -> tp.Union[T, tp.Iterator[T]]:
         return self.items[item]
 
     def remove(self, other: T) -> None:
@@ -77,3 +79,49 @@ class SortedList(tp.Generic[T]):
         self.keys.insert(index, key_value)
 
         return index
+
+
+class SliceableDeque:
+    """
+    A collections.deque that supports slicing
+    """
+    __slots__ = ('deque', )
+
+    def __bool__(self) -> bool:
+        return bool(self.deque)
+
+    def __init__(self, *args, **kwargs):
+        self.deque = collections.deque(*args, **kwargs)
+
+    def __setitem__(self, key: int, value: T) -> None:
+        self.deque[key] = value
+
+    def __delitem__(self, key: int) -> None:
+        del self.deque[key]
+
+    def __iter__(self) -> tp.Iterator[T]:
+        return iter(self.deque)
+
+    def __len__(self) -> int:
+        return len(self.deque)
+
+    def __reversed__(self) -> tp.Iterator[T]:
+        return reversed(self.deque)
+
+    def __contains__(self, item: T) -> bool:
+        return item in self.deque
+
+    def __getattr__(self, item: str):
+        return getattr(self.deque, item)
+
+    def __getitem__(self, item) -> tp.Union[tp.Iterator[T], T]:
+        tot_length = len(self)
+        if type(item) is slice:
+            start, stop, step = item.indices(tot_length)
+
+            def generator():
+                for index in range(start, stop, step):
+                    yield self.deque[index]
+            return generator()
+        else:
+            return self.deque[item]